Split big test module up into multiple (more concurrency)
